Safari-611.1.10/LayoutTests/fast/events/event-leak-objects-expected.txt


Testing ErrorEvent.error
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.

Testing PopStateEvent.state
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.

Testing MessageEvent.data
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.

Testing CustomEvent.detail
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS window.internals.isFromCurrentWorld(event) is true
PASS checkAllPropertiesFromCurrentWorld(resultValue) is true
PASS Object.pageDefinedVar is undefined.
PASS successfullyParsed is true

TEST COMPLETE
